Changed notation for string from char* to char[] + unsigned int -> size_t

+++ ft_strmapi.c
@@ -6,17 +6,19 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 19:05:00 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/10 18:47:15 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/11 17:09:12 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include <stddef.h>
 
 #include "libft.h"
 
 char	*ft_strmapi(char const s[], char (*f)(unsigned int, char))
 {
-	char			*str;
-	unsigned int	i;
-	unsigned int	len;
+	char	*str;
+	size_t	i;
+	size_t	len;
 
 	if (s == 0 || f == 0)
 		return (0);
